Start/continue new/old repository
This repo is remainder of my previous repo. Why is that? So, when I tried create new branch and remove useless commits, I broke my repo and return my code to old version)) So I had to rewrite my project a little bit and create new repo for this. In this commit:
1. Fixed my cookies and sessions at entry
2. Did on 50% search line on admin page